Revert "chore: add all controllers documentation"
