+++ src/leaderboards/leaderboards.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Get, Inject, QueryParams } from '@tsed/common';
-import { ContentType, Description, Hidden, Returns, Summary } from '@tsed/schema';
+import { ContentType, Hidden } from '@tsed/schema';
 import { CachedBoost } from './interface/cached-boost.interface';
-import { LeaderBoardDataModel } from './interface/leaderboard-data-model.interrface';
+import { LeaderBoardData } from './interface/leaderboard-data.interrface';
 import { LeaderBoardsService } from './leaderboards.service';
 
 @Controller('/leaderboards')
@@ -16,15 +16,12 @@ export class LeaderBoardsController {
     return this.leaderBoardsService.loadFullLeaderBoard();
   }
 
-  @Get('')
   @ContentType('json')
-  @Summary('Get a page of the badger boost leaderboard')
-  @Description('Return the requested page of the boost leaderboard')
-  @Returns(200, LeaderBoardDataModel)
+  @Get('')
   async getLeaderBoard(
     @QueryParams('page') page?: number,
     @QueryParams('size') size?: number,
-  ): Promise<LeaderBoardDataModel> {
+  ): Promise<LeaderBoardData> {
     return this.leaderBoardsService.loadLeaderboardEntries(page, size);
   }
 }
