+++ src/prices/prices.controller.ts
@@ -1,9 +1,8 @@
 import { Controller, Get, Inject, QueryParams } from '@tsed/common';
-import { ContentType, Description, Returns, Summary } from '@tsed/schema';
+import { ContentType } from '@tsed/schema';
 import { Chain } from '../chains/config/chain.config';
 import { ChainNetwork } from '../chains/enums/chain-network.enum';
-import { PriceSummary } from '../tokens/interfaces/price-summary.interface';
-import { PriceSummaryModel } from '../tokens/interfaces/price-summary-model.interface';
+import { PriceSummary } from '../tokens/interfaces/token-price.interface';
 import { PricesService } from './prices.service';
 
 @Controller('/prices')
@@ -11,11 +10,8 @@ export class PriceController {
   @Inject()
   pricesService!: PricesService;
 
-  @Get('')
   @ContentType('json')
-  @Summary('Get a summary of token prices related to the Badger Protocol')
-  @Description('Return a map of checksum contract address to the currency value of the token')
-  @Returns(200, PriceSummaryModel)
+  @Get('/')
   async listPrices(
     @QueryParams('chain') chain?: ChainNetwork,
     @QueryParams('currency') currency?: string,
