+++ src/prices/prices.utils.ts
@@ -10,10 +10,9 @@ import { ibBTCAbi } from '../config/abi/ibbtc.abi';
 import { COINGECKO_URL, TOKENS } from '../config/constants';
 import { getSett } from '../setts/setts.utils';
 import { TokenType } from '../tokens/enums/token-type.enum';
-import { PriceData } from '../tokens/interfaces/price-data.interface';
 import { Token } from '../tokens/interfaces/token.interface';
 import { TokenConfig } from '../tokens/interfaces/token-config.interface';
-import { TokenPrice } from '../tokens/interfaces/token-price.interface';
+import { PriceData, TokenPrice } from '../tokens/interfaces/token-price.interface';
 import { TokenPriceSnapshot } from '../tokens/interfaces/token-price-snapshot.interface';
 import { getToken, getTokenByName } from '../tokens/tokens.utils';
