--- src/rewards/interfaces/reward-merkle-claim.interface.ts
@@ -1,12 +1,0 @@
-import { BigNumber } from 'ethers';
-
-export interface RewardMerkleClaim {
-  index: BigNumber;
-  cycle: BigNumber;
-  boost: BigNumber;
-  user: string;
-  tokens: string[];
-  cumulativeAmounts: BigNumber[];
-  proof: string[];
-  node: string;
-}
