--- src/tokens/interfaces/price-summary.interface.ts
@@ -1,7 +1,0 @@
-/**
- * Mapping from token contract address to single
- * currency price data.
- */
-export interface PriceSummary {
-  [address: string]: number;
-}
