+++ src/tokens/interfaces/token-price.interface.ts
@@ -8,3 +8,19 @@ export interface TokenPrice {
   usd: number;
   eth: number;
 }
+
+/**
+ * Mapping from token contract address to detailed
+ * token price data.
+ */
+export interface PriceData {
+  [address: string]: TokenPrice;
+}
+
+/**
+ * Mapping from token contract address to single
+ * currency price data.
+ */
+export interface PriceSummary {
+  [address: string]: number;
+}
